removed home page
