+++ src/components/bread-crumb-nav/bread-crumb-nav.test.js
@@ -5,63 +5,20 @@ import BreadCrumbNav from './bread-crumb-nav';
 import { PATHS } from 'common';
 
 describe('BreadCrumbNav', () => {
-  it('should render', () => {
+  it('should render - create', async () => {
     render(
       <MemoryRouter initialEntries={[PATHS.INDEX]}>
         <BreadCrumbNav />
       </MemoryRouter>
     );
 
-    // Home
-    let homeButton = screen.getAllByText('home')[0];
-    let createButtons = screen.queryAllByText('create');
-    expect(homeButton).toBeInTheDocument();
-    expect(createButtons).toEqual([]);
-
-    // Redirect test (itself)
-    fireEvent.click(homeButton);
-
-    homeButton = screen.getAllByText('home')[0];
-    createButtons = screen.queryAllByText('create');
-    expect(homeButton).toBeInTheDocument();
-    expect(createButtons).toEqual([]);
-  });
-
-  it('should render - create', async () => {
-    render(
-      <MemoryRouter initialEntries={[PATHS.CREATE_MANIFEST]}>
-        <BreadCrumbNav />
-      </MemoryRouter>
-    );
-
     // Home > Create
-    let homeButton = screen.getAllByText('home')[0];
-    let createButtons = screen.getAllByText('create')[0];
+    const homeButton = screen.getAllByText('home')[0];
     expect(homeButton).toBeInTheDocument();
-    expect(createButtons).toBeInTheDocument();
-
-    // Redirect test (itself)
-    fireEvent.click(createButtons);
-
-    homeButton = screen.getAllByText('home')[0];
-    createButtons = screen.getAllByText('create')[0];
-    expect(homeButton).toBeInTheDocument();
-    expect(createButtons).toBeInTheDocument();
 
     // Redirect test (Home)
     fireEvent.click(homeButton);
 
-    homeButton = screen.getAllByText('home')[0];
-    createButtons = screen.queryAllByText('create');
-    expect(homeButton).toBeInTheDocument();
-    expect(createButtons).toEqual([]);
-  });
-
-  it('should match snapshot - edit', () => {
-    expect(render(
-      <MemoryRouter initialEntries={[PATHS.EDIT_MANIFEST]}>
-        <BreadCrumbNav />
-      </MemoryRouter>)
-    ).toMatchSnapshot();
+    expect(screen.getAllByText('home')[0]).toBeInTheDocument();
   });
 });
