+++ src/pages/create-manifest/create-manifest.js
@@ -1,5 +1,4 @@
 import { useState, useCallback, useEffect, useMemo } from 'react';
-import PropTypes from 'prop-types';
 import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
 import { cx } from '@emotion/css';
@@ -9,7 +8,7 @@ import { shallow } from 'zustand/shallow';
 import FieldLabel from 'components/field-label';
 import { PATHS } from 'common';
 import { getEnvs } from 'common/functions';
-import { useUserStore, useResponseStore, useReviewManifestStore } from 'common/store';
+import { useUserStore, useResponseStore } from 'common/store';
 import FileField from './file-field';
 import {
   buttonLabelStyle,
@@ -39,20 +38,17 @@ export const TEST_ID = {
 };
 
 const userStoreSelector = (s) => [s.setGeography, s.geography, s.setSubscriptionKey, s.subscriptionKey];
-const responseStoreSelector = (s) => [s.acknowledgeError, s.errorMessage, s.uploadFile, s.setExistingManifestJson];
-const reviewManifestSelector = (s) => s.setOriginalPackage;
+const responseStoreSelector = (s) => [s.acknowledgeError, s.errorMessage, s.uploadFile];
 
-const CreateManifestPage = ({ allowEdit }) => {
+const CreateManifestPage = () => {
   const { t } = useTranslation();
   const navigate = useNavigate();
 
   const [errorMessage, setErrorMessage] = useState('');
   const [file, setFile] = useState(null);
-  const [manifestFileChosen, setManifestFileChosen] = useState(false);
 
   const [setGeo, geo, setSubKey, subKey] = useUserStore(userStoreSelector, shallow);
-  const [acknowledgeApiError, apiErrorMessage, uploadFile, setExistingManifestJson] = useResponseStore(responseStoreSelector, shallow);
-  const setOriginalPackage = useReviewManifestStore(reviewManifestSelector);
+  const [acknowledgeApiError, apiErrorMessage, uploadFile] = useResponseStore(responseStoreSelector, shallow);
 
   const environmentOptions = useMemo(() => (
     Object.keys(getEnvs()).map((geography) => ({
@@ -61,8 +57,7 @@ const CreateManifestPage = ({ allowEdit }) => {
     }))
   ), [t]);
 
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  useEffect(() => acknowledgeApiError, []);
+  useEffect(() => acknowledgeApiError, []); // eslint-disable-line react-hooks/exhaustive-deps
 
   useEffect(() => {
     if (apiErrorMessage) {
@@ -75,36 +70,17 @@ const CreateManifestPage = ({ allowEdit }) => {
       return;
     }
 
-    if (!allowEdit) {
-      setExistingManifestJson(null);
-    }
-
     uploadFile(file);
-    setOriginalPackage(file);
     navigate(PATHS.PROCESSING);
-  }, [allowEdit, file, navigate, subKey, setExistingManifestJson, uploadFile, setOriginalPackage]);
+  }, [file, navigate, subKey, uploadFile]);
 
   const navigateHome = useCallback(() => navigate(PATHS.INDEX), [navigate]);
   const updateSubKey = useCallback((e) => setSubKey(e.target.value), [setSubKey]);
   const updateGeo = useCallback((_, option) => setGeo(option.key), [setGeo]);
 
   const allFieldsFilled = useMemo(() => {
-    const createManifestFieldsFilled = file !== null && subKey !== '';
-    if (allowEdit) {
-      return manifestFileChosen && createManifestFieldsFilled;
-    }
-    return createManifestFieldsFilled;
-  }, [allowEdit, file, manifestFileChosen, subKey]);
-
-  const saveManifestFileData = (file) => {
-    new Response(file).json().then(json => {
-      setManifestFileChosen(true);
-      setExistingManifestJson(json);
-    }, () => {
-      setManifestFileChosen(false);
-      setExistingManifestJson(null);
-    });
-  };
+    return file !== null && subKey !== '';
+  }, [file, subKey]);
 
   return (
     <div className={containerStyle}>
@@ -133,23 +109,12 @@ const CreateManifestPage = ({ allowEdit }) => {
       </div>
       <FileField label={t('dwg.zip.package')} id={TEST_ID.FILE_UPLOAD_FIELD} onFileSelect={setFile}
                  fileType='zip' onError={setErrorMessage} tooltip={t('tooltip.dwg.zip.package')} />
-      {allowEdit && <FileField label={t('manifest.file')} id={TEST_ID.MANIFEST_UPLOAD_FIELD}
-                               onFileSelect={saveManifestFileData} fileType='json'
-                               tooltip={t('tooltip.manifest.file')} onError={setErrorMessage} />}
       <PrimaryButton disabled={!allFieldsFilled} onClick={uploadButtonOnClick} data-testid={TEST_ID.UPLOAD_BUTTON}
                      className={primaryButtonStyle} text={t('process')} styles={primaryButtonDisabledStyles} />
       <DefaultButton styles={buttonLabelStyle} className={defaultButtonStyle} text={t('cancel')}
                      onClick={navigateHome} data-testid={TEST_ID.CANCEL_BUTTON} />
     </div>
   );
-};
-
-CreateManifestPage.propTypes = {
-  allowEdit: PropTypes.bool,
-};
-
-CreateManifestPage.defaultProps = {
-  allowEdit: false,
 };
 
 export default CreateManifestPage;
