+++ src/pages/create-manifest/create-manifest.test.js
@@ -6,7 +6,6 @@ import { uploadFile } from 'common/api';
 import { useResponseStore } from 'common/store';
 
 const mockNavigate = jest.fn();
-const mockSetOriginalPackage = jest.fn();
 
 jest.mock('react-router-dom', () => ({
   useNavigate: () => mockNavigate,
@@ -15,16 +14,11 @@ jest.mock('common/api', () => ({
   ...jest.requireActual('common/api'),
   uploadFile: jest.fn(),
 }));
-jest.mock('common/store/review-manifest.store', () => ({
-  useReviewManifestStore: () => mockSetOriginalPackage,
-}));
 
 describe('CreateManifestPage', () => {
   let file;
-  let manifestJsonFile;
   const state = useResponseStore.getState();
   const uploadFileSpy = jest.spyOn(state, 'uploadFile');
-  const setExistingManifestJsonSpy = jest.spyOn(state, 'setExistingManifestJson');
 
   beforeEach(() => {
     uploadFile.mockReturnValue(Promise.resolve({}));
@@ -33,10 +27,6 @@ describe('CreateManifestPage', () => {
       size: 1024 * 1024 * 10,
       type: 'application/x-zip-compressed',
     };
-    manifestJsonFile = new File(['{"foo":"bar"}'], 'manifest.json', {
-      type: 'application/json',
-      lastModified: Date.now(),
-    });
     useResponseStore.setState({
       errorMessage: 'some err',
     });
@@ -48,15 +38,8 @@ describe('CreateManifestPage', () => {
     expect(view).toMatchSnapshot();
   });
 
-  it('should render correctly with allowEdit prop', () => {
-    const view = render(<CreateManifestPage allowEdit />);
-    expect(view).toMatchSnapshot();
-  });
-
   it('should upload file correctly', () => {
     expect(uploadFileSpy).not.toHaveBeenCalled();
-    expect(mockSetOriginalPackage).not.toHaveBeenCalled();
-    expect(setExistingManifestJsonSpy).not.toHaveBeenCalled();
     render(<CreateManifestPage />);
     const fileInput = screen.getByTestId(TEST_ID.FILE_UPLOAD_FIELD);
     const uploadButton = screen.getByTestId(TEST_ID.UPLOAD_BUTTON);
@@ -78,45 +61,6 @@ describe('CreateManifestPage', () => {
     fireEvent.click(uploadButton);
 
     expect(uploadFileSpy).toBeCalledWith(file);
-    expect(setExistingManifestJsonSpy).toBeCalledWith(null);
-    expect(mockSetOriginalPackage).toHaveBeenCalledWith(file);
-  });
-
-  it('should upload file correctly with allowEdit true', async () => {
-    expect(uploadFileSpy).not.toHaveBeenCalled();
-    expect(mockSetOriginalPackage).not.toHaveBeenCalled();
-    expect(setExistingManifestJsonSpy).not.toHaveBeenCalled();
-    render(<CreateManifestPage allowEdit />);
-    const fileInput = screen.getByTestId(TEST_ID.FILE_UPLOAD_FIELD);
-    const manifestJsonInput = screen.getByTestId(TEST_ID.MANIFEST_UPLOAD_FIELD);
-    const uploadButton = screen.getByTestId(TEST_ID.UPLOAD_BUTTON);
-    const subKeyTextField = screen.getByTestId(TEST_ID.SUBSCRIPTION_KEY_FIELD);
-
-    fireEvent.change(fileInput, {
-      target: { files: [file] },
-    });
-    fireEvent.change(manifestJsonInput, {
-      target: { files: [manifestJsonFile] },
-    });
-
-    expect(screen.getAllByTestId(TEST_ID.FILE_NAME_FIELD)[0].value).toBe(file.name);
-    expect(screen.getAllByTestId(TEST_ID.FILE_NAME_FIELD)[1].value).toBe(manifestJsonFile.name);
-    expect(screen.queryByText('error.file.size.exceeded')).toBeNull();
-    expect(screen.queryByText('error.file.type.incorrect')).toBeNull();
-
-    fireEvent.change(subKeyTextField, {
-      target: {
-        value: 'Sonic the Hedgehog isnt his fullname',
-      },
-    });
-
-    await waitFor(() => expect(uploadButton).not.toBeDisabled());
-
-    fireEvent.click(uploadButton);
-
-    expect(uploadFileSpy).toBeCalledWith(file);
-    expect(setExistingManifestJsonSpy).toBeCalledWith({ foo: 'bar' });
-    expect(mockSetOriginalPackage).toHaveBeenCalledWith(file);
   });
 
   it('should show an error when the file is too big',  async () => {
