--- src/pages/create-manifest/edit-manifest.js
@@ -1,5 +1,0 @@
-import CreateManifest from './create-manifest';
-
-const EditManifest = () => <CreateManifest allowEdit />;
-
-export default EditManifest;
