--- src/pages/index/index.style.js
@@ -1,20 +1,0 @@
-import { css } from '@emotion/css';
-
-export const containerStyle = css`
-  margin-top: 3.375rem;
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-`;
-
-export const mainDescriptionStyle = css`
-  max-width: 560px;
-  text-align: center;
-`;
-
-export const paneSectionStyle = css`
-  display: flex;
-  flex-flow: row wrap;
-  gap: 1rem;
-  margin: 1rem 0;
-`;
